Modularize RCCL, reduce std logs, define clear output contract
