+++ cvs/lib/rccl/__init__.py
@@ -1,0 +1,1 @@
+"""Internal RCCL helpers split out of `cvs.lib.rccl_cvs`."""
